HBASE-29756: Programmatically register related co-processor during initialization (#7743)
* HBASE-29756: Programmatically register related co-processor during initialization

* Apply Spotless

* Remove the cached globalReadOnlyMode variable and make manageclusterIDFile static

* Address review comments

* Address review comments

* Make coprocessor addition and removal generic

* Make manageClusterIdFile Idempotent

* Address review comments

* Avoid intelliJ warning about fixed size array creation

+++ hbase-server/src/main/java/org/apache/hadoop/hbase/security/access/AbstractReadOnlyController.java
@@ -18,7 +18,6 @@
 package org.apache.hadoop.hbase.security.access;
 
 import java.io.IOException;
-import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.fs.FileSystem;
 import org.apache.hadoop.fs.Path;
 import org.apache.hadoop.hbase.Coprocessor;
@@ -26,8 +25,6 @@ import org.apache.hadoop.hbase.CoprocessorEnvironment;
 import org.apache.hadoop.hbase.DoNotRetryIOException;
 import org.apache.hadoop.hbase.HBaseInterfaceAudience;
 import org.apache.hadoop.hbase.HConstants;
-import org.apache.hadoop.hbase.conf.ConfigurationObserver;
-import org.apache.hadoop.hbase.coprocessor.MasterCoprocessorEnvironment;
 import org.apache.hadoop.hbase.master.MasterFileSystem;
 import org.apache.hadoop.hbase.master.MasterServices;
 import org.apache.hadoop.hbase.util.FSUtils;
@@ -35,40 +32,24 @@ import org.apache.yetus.audience.InterfaceAudience;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-
 @InterfaceAudience.LimitedPrivate(HBaseInterfaceAudience.CONFIG)
-public abstract class AbstractReadOnlyController implements ConfigurationObserver, Coprocessor {
-  protected volatile boolean globalReadOnlyEnabled;
+public abstract class AbstractReadOnlyController implements Coprocessor {
   private MasterServices masterServices;
   private static final Logger LOG = LoggerFactory.getLogger(AbstractReadOnlyController.class);
 
   protected void internalReadOnlyGuard() throws DoNotRetryIOException {
-    if (this.globalReadOnlyEnabled) {
-      throw new DoNotRetryIOException("Operation not allowed in Read-Only Mode");
-    }
+    throw new DoNotRetryIOException("Operation not allowed in Read-Only Mode");
   }
 
   @Override
   public void start(CoprocessorEnvironment env) throws IOException {
-    if (env instanceof MasterCoprocessorEnvironment) {
-      this.masterServices = ((MasterCoprocessorEnvironment) env).getMasterServices();
-      LOG.info("ReadOnlyController obtained MasterServices reference from start().");
-    } else {
-      LOG.debug("ReadOnlyController loaded in a non-Master environment. "
-        + "File system operations for read-only state will not work.");
-    }
-
-    this.globalReadOnlyEnabled =
-      env.getConfiguration().getBoolean(HConstants.HBASE_GLOBAL_READONLY_ENABLED_KEY,
-        HConstants.HBASE_GLOBAL_READONLY_ENABLED_DEFAULT);
   }
 
   @Override
   public void stop(CoprocessorEnvironment env) {
   }
 
-  private void manageActiveClusterIdFile(boolean readOnlyEnabled) {
-    MasterFileSystem mfs = this.masterServices.getMasterFileSystem();
+  public static void manageActiveClusterIdFile(boolean readOnlyEnabled, MasterFileSystem mfs) {
     FileSystem fs = mfs.getFileSystem();
     Path rootDir = mfs.getRootDir();
     Path activeClusterFile = new Path(rootDir, HConstants.ACTIVE_CLUSTER_SUFFIX_FILE_NAME);
@@ -79,8 +60,13 @@ public abstract class AbstractReadOnlyController implements ConfigurationObserve
         LOG.debug("Global read-only mode is being ENABLED. Deleting active cluster file: {}",
           activeClusterFile);
         try {
-          fs.delete(activeClusterFile, false);
-          LOG.info("Successfully deleted active cluster file: {}", activeClusterFile);
+          if (fs.exists(activeClusterFile)) {
+            fs.delete(activeClusterFile, false);
+            LOG.info("Successfully deleted active cluster file: {}", activeClusterFile);
+          } else {
+            LOG.debug("Active cluster file does not exist at: {}. No need to delete.",
+              activeClusterFile);
+          }
         } catch (IOException e) {
           LOG.error(
             "Failed to delete active cluster file: {}. "
@@ -90,7 +76,12 @@ public abstract class AbstractReadOnlyController implements ConfigurationObserve
       } else {
         // DISABLING READ-ONLY (true -> false), create the active cluster file id file
         int wait = mfs.getConfiguration().getInt(HConstants.THREAD_WAKE_FREQUENCY, 10 * 1000);
-        FSUtils.setActiveClusterSuffix(fs, rootDir, mfs.getSuffixFileDataToWrite(), wait);
+        if (!fs.exists(activeClusterFile)) {
+          FSUtils.setActiveClusterSuffix(fs, rootDir, mfs.getSuffixFileDataToWrite(), wait);
+        } else {
+          LOG.debug("Active cluster file already exists at: {}. No need to create it again.",
+            activeClusterFile);
+        }
       }
     } catch (IOException e) {
       // We still update the flag, but log that the operation failed.
@@ -98,22 +89,4 @@ public abstract class AbstractReadOnlyController implements ConfigurationObserve
         + "Flag will be updated, but file system state may be inconsistent.", e);
     }
   }
-
-  @Override
-  public void onConfigurationChange(Configuration conf) {
-    boolean maybeUpdatedConfValue = conf.getBoolean(HConstants.HBASE_GLOBAL_READONLY_ENABLED_KEY,
-      HConstants.HBASE_GLOBAL_READONLY_ENABLED_DEFAULT);
-    if (this.globalReadOnlyEnabled != maybeUpdatedConfValue) {
-      if (this.masterServices != null) {
-        manageActiveClusterIdFile(maybeUpdatedConfValue);
-      } else {
-        LOG.debug("Global R/O flag changed, but not running on master");
-      }
-
-      this.globalReadOnlyEnabled = maybeUpdatedConfValue;
-      LOG.info("Config {} has been dynamically changed to {}.",
-        HConstants.HBASE_GLOBAL_READONLY_ENABLED_KEY, this.globalReadOnlyEnabled);
-    }
-  }
-
 }
